+++ ccdproc/core.py
@@ -2,10 +2,13 @@
 
 """This module implements the base CCDPROC functions"""
 
+import math
 import numbers
+import logging
 
 import numpy as np
-import math
+from scipy import ndimage
+
 from astropy.units.quantity import Quantity
 from astropy import units as u
 from astropy.modeling import fitting
@@ -16,11 +19,11 @@ from astropy.wcs.utils import proj_plane_pixel_area
 from astropy.utils import deprecated
 import astropy  # To get the version.
 
-from scipy import ndimage
-
 from .utils.slices import slice_from_string
 from .log_meta import log_to_metadata
 from .extern.bitfield import bitfield_to_boolean_mask as _bitfield_to_boolean_mask
+
+logger = logging.getLogger(__name__)
 
 __all__ = ['background_deviation_box', 'background_deviation_filter',
            'ccd_process', 'cosmicray_median', 'cosmicray_lacosmic',
@@ -270,7 +273,7 @@ def ccd_process(ccd, oscan=None, trim=None, error=False, master_bias=None,
 
 
 @log_to_metadata
-def create_deviation(ccd_data, gain=None, readnoise=None):
+def create_deviation(ccd_data, gain=None, readnoise=None, disregard_nan=False):
     """
     Create a uncertainty frame. The function will update the uncertainty
     plane which gives the standard deviation for the data. Gain is used in
@@ -292,6 +295,10 @@ def create_deviation(ccd_data, gain=None, readnoise=None):
     readnoise : `~astropy.units.Quantity` or None, optional
         Read noise per pixel.
         Default is ``None``.
+
+    disregard_nan: boolean
+        If ``True``, any value of nan in the output array will be replaced by
+        the readnoise.
 
     {log}
 
@@ -328,9 +335,20 @@ def create_deviation(ccd_data, gain=None, readnoise=None):
     gain_value = float(gain / gain.unit)
     readnoise_value = float(readnoise / readnoise.unit)
 
-    var = (gain_value * ccd_data.data + readnoise_value ** 2) ** 0.5
-    ccd = ccd_data.copy()
+    # remove values that might be negative or treat as nan
+    data = gain_value * ccd_data.data
+    mask = (data < 0)
+    if disregard_nan:
+        data[mask] = 0
+    else:
+        data[mask] = np.nan
+        logging.warning('Negative values in array will be replaced with nan')
+
+    # calculate the deviation
+    var = (data + readnoise_value ** 2) ** 0.5
+
     # ensure uncertainty and image data have same unit
+    ccd = ccd_data.copy()
     var /= gain_value
     ccd.uncertainty = StdDevUncertainty(var)
     return ccd
@@ -617,6 +635,10 @@ def subtract_dark(ccd, master, dark_exposure=None, data_exposure=None,
     result : `~astropy.nddata.CCDData`
         Dark-subtracted image.
     """
+    if ccd.shape != master.shape:
+        err_str = "operands could not be subtracted with shapes {} {}".format(ccd.shape, master.shape)
+        raise ValueError(err_str)
+
     if not (isinstance(ccd, CCDData) and isinstance(master, CCDData)):
         raise TypeError("ccd and master must both be CCDData objects.")
 
@@ -849,6 +871,10 @@ def transform_image(ccd, transform_func, **kwargs):
         mask = transform_func(nccd.mask, **kwargs)
         nccd.mask = mask > 0
 
+    if nccd.wcs is not None:
+        warn = 'WCS information may be incorrect as no transformation was applied to it'
+        logging.warning(warn)
+
     return nccd
 
 
@@ -891,6 +917,7 @@ def wcs_project(ccd, target_wcs, target_shape=None, order='bilinear'):
     ccd : `~astropy.nddata.CCDData`
         A transformed CCDData object.
     """
+    from astropy.nddata.ccddata import _generate_wcs_and_update_header
     from reproject import reproject_interp
 
     if not (ccd.wcs.is_celestial and target_wcs.is_celestial):
@@ -930,9 +957,12 @@ def wcs_project(ccd, target_wcs, target_shape=None, order='bilinear'):
     if not output_mask.any():
         output_mask = None
 
+    # If there are any wcs keywords in the header, remove them
+    hdr, _ = _generate_wcs_and_update_header(ccd.header)
+
     nccd = CCDData(area_ratio * projected_image_raw, wcs=target_wcs,
                    mask=output_mask,
-                   header=ccd.header, unit=ccd.unit)
+                   header=hdr, unit=ccd.unit)
 
     return nccd
 
@@ -1857,7 +1887,7 @@ def bitfield_to_boolean_mask(bitfield, ignore_bits=0, flip_bits=None):
         good_mask_value=False, dtype=bool)
 
 
-class Keyword(object):
+class Keyword:
     """
     """
     def __init__(self, name, unit=None, value=None):